requirements.txt
pandas==1.1.5
numpy==1.21.0
instaloader==4.8.4
ipykernel==6.7.0
missingno==0.5.0
nose==1.3.7
autopep8==1.6.0
yapf==0.32.0
scikit-learn==1.0.2
wandb==0.12.10
statsmodels==0.13.1
jupyter-dash==0.4.1
plotly==5.6.0
darts==0.17.1
imblearn==0.0
dash-bootstrap-components==1.0.3
https://github.com/aboSamoor/pycld2/zipball/e3ac86ed4d4902e912691c1531d0c5645382a726 # for detecting language